MAINT: special: remove _spence.pxd and use xsf instead
