+++ scipy/special/meson.build
@@ -7,7 +7,6 @@ _ufuncs_pxi_pxd_sources = [
   fs.copyfile('_hyp0f1.pxd'),
   fs.copyfile('_hypergeometric.pxd'),
   fs.copyfile('_legacy.pxd'),
-  fs.copyfile('_spence.pxd'),
   fs.copyfile('orthogonal_eval.pxd'),
   fs.copyfile('sf_error.pxd'),
   fs.copyfile('_ufuncs_extra_code.pxi'),
